refactor(codegen): remove unnecessary program counters
